Add alert views

+++ wallEEt/New Group/ExpensesModel.swift
@@ -10,11 +10,11 @@ import Foundation
 import RealmSwift
 class ExpensesModel:Object,AccountItem{
     //Properties
-    dynamic var amount: Int = 0
+    dynamic var amount: Double = 0
     dynamic var date: Date!
     dynamic var comment: String = ""
     //Init
-    convenience init(amount:Int, date: Date, comment:String){
+    convenience init(amount:Double, date: Date, comment:String){
         self.init()
         self.amount = amount
         self.date = date
