+++ wallEEt/New Group/IncomesModel.swift
@@ -13,11 +13,11 @@ enum IncomeSources:String{
 }
 class IncomesModel:Object,AccountItem{
     //Properties
-    dynamic var amount:Int = 0
+    dynamic var amount:Double = 0
     dynamic var date:Date!
     dynamic var source:String = ""
     //Init
-    convenience init(amount:Int,date:Date,source:IncomeSources){
+    convenience init(amount:Double,date:Date,source:IncomeSources){
         self.init()
         self.amount = amount
         self.date = date
